requirements.txt
fastapi
uvicorn[standard]
gradio-client
groq
httpx
python-dotenv
sqlalchemy
psycopg2-binary
gunicorn
random
